requirements.txt
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.0.3
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
PyJWT==2.9.0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.4
requests==2.32.3
simple-websocket==1.0.0
urllib3==2.2.3
Werkzeug==3.0.4
wsproto==1.2.0
